requirements_dev.txt
torch
pytorch-lightning
optuna
optuna-dashboard
tensorboard

numpy
pandas
matplotlib
seaborn

rdkit
biopython

jupyter
tqdm
black

